fix(backend): return correct total count

+++ crates/models/dependent/src/lib.rs
@@ -202,7 +202,7 @@ pub enum DailyUserActivitiesResponseGroupedBy {
 #[derive(Debug, Serialize, Deserialize, SimpleObject, Clone)]
 pub struct DailyUserActivitiesResponse {
     pub items: Vec<daily_user_activity::Model>,
-    pub total_count: usize,
+    pub total_count: i64,
     pub most_active_hour: u8,
     pub grouped_by: DailyUserActivitiesResponseGroupedBy,
 }
